% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compare_all}
\alias{compare_all}
\title{Runtime comparison}
\format{much bigger data set
\itemize{
  \item list all variables
}}
\usage{
compare_all
}
\description{
Runtime comparison
}
\keyword{datasets}

