% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{runtimes_protein}
\alias{runtimes_protein}
\title{Runtime results from protein example}
\format{a data frame consisting of 100 comparison runs across five different packages
\itemize{
  \item iteration iteration number
  \item igraph runtime under the igraph package
}}
\usage{
runtimes_protein
}
\description{
Runtime results from protein example
}
\keyword{datasets}

